Move HiddenUsersState to commons/commonMain, nip03Timestamp to commons/jvmAndroid
- HiddenUsersState: moved to commons/commonMain, replaced Dispatchers.IO with Dispatchers_IO for iOS compat
- nip03Timestamp (4 files): moved to commons/jvmAndroid
  - OtsState: LocalCache -> ICacheProvider
  - IncomingOtsEventVerifier: LocalCache -> ICacheProvider (uses getEventStream())
  - OtsSettings: moved to jvmAndroid (depends on okhttp3 OkHttpBitcoinExplorer)
  - TorAwareOkHttpOtsResolverBuilder: moved to jvmAndroid (okhttp3)
- Fixed smart cast issue in OtsSharedPreferences (cross-module property)
- All original files replaced with typealiases for backward compat
