+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip03Timestamp/OtsState.kt
@@ -20,85 +20,5 @@
  */
 package com.vitorpamplona.amethyst.model.nip03Timestamp
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.quartz.nip01Core.core.Event
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip03Timestamp.OtsEvent
-import com.vitorpamplona.quartz.nip03Timestamp.OtsResolver
-import com.vitorpamplona.quartz.utils.Log
-import com.vitorpamplona.quartz.utils.TimeUtils
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.launch
-import java.util.Base64
-
-class OtsState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val otsResolver: () -> OtsResolver,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    var lastTimeItTriedToUpdateAttestations: Long = 0
-
-    fun upgradeAttestationsIfNeeded(onReady: (List<OtsEvent>) -> Unit) {
-        // only tries to upgrade every hour
-        val now = TimeUtils.now()
-        if (now - lastTimeItTriedToUpdateAttestations > TimeUtils.ONE_HOUR) {
-            lastTimeItTriedToUpdateAttestations = now
-            scope.launch {
-                onReady(updateAttestations())
-            }
-        }
-    }
-
-    suspend fun updateAttestations(): List<OtsEvent> {
-        Log.d("Pending Attestations") { "Updating ${settings.pendingAttestations.value.size} pending attestations" }
-
-        return settings.pendingAttestations.value.toList().mapNotNull { (key, value) ->
-            val otsState = OtsEvent.upgrade(Base64.getDecoder().decode(value), key, otsResolver())
-
-            if (otsState != null) {
-                val hint = cache.getNoteIfExists(key)?.toEventHint<Event>()
-                val template =
-                    if (hint != null) {
-                        OtsEvent.build(hint, otsState)
-                    } else {
-                        OtsEvent.build(key, otsState)
-                    }
-
-                val otsEvent = signer.sign(template)
-
-                settings.pendingAttestations.update { it - key }
-
-                otsEvent
-            } else {
-                null
-            }
-        }
-    }
-
-    fun hasPendingAttestations(note: Note): Boolean {
-        val id = note.event?.id ?: note.idHex
-        return settings.pendingAttestations.value[id] != null
-    }
-
-    suspend fun timestamp(note: Note) {
-        if (note.isDraft()) return
-
-        val id = note.event?.id ?: note.idHex
-
-        settings.addPendingAttestation(
-            id = id,
-            stamp =
-                Base64.getEncoder().encodeToString(
-                    OtsEvent.stamp(
-                        id,
-                        otsResolver(),
-                    ),
-                ),
-        )
-    }
-}
+// Re-export from commons for backward compatibility
+typealias OtsState = com.vitorpamplona.amethyst.commons.model.nip03Timestamp.OtsState
