+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip03Timestamp/TorAwareOkHttpOtsResolverBuilder.kt
@@ -20,37 +20,5 @@
  */
 package com.vitorpamplona.amethyst.model.nip03Timestamp
 
-import com.vitorpamplona.quartz.nip03Timestamp.OtsResolver
-import com.vitorpamplona.quartz.nip03Timestamp.OtsResolverBuilder
-import com.vitorpamplona.quartz.nip03Timestamp.okhttp.OkHttpBitcoinExplorer
-import com.vitorpamplona.quartz.nip03Timestamp.okhttp.OkHttpCalendar
-import com.vitorpamplona.quartz.nip03Timestamp.ots.OtsBlockHeightCache
-import okhttp3.OkHttpClient
-
-class TorAwareOkHttpOtsResolverBuilder(
-    val okHttpClient: (url: String) -> OkHttpClient,
-    val isTorActive: (url: String) -> Boolean,
-    val cache: OtsBlockHeightCache,
-    val customExplorerUrl: () -> String? = { null },
-) : OtsResolverBuilder {
-    fun getAPI(usingTor: Boolean): String =
-        customExplorerUrl()
-            ?: if (usingTor) {
-                OkHttpBitcoinExplorer.MEMPOOL_API_URL
-            } else {
-                OkHttpBitcoinExplorer.BLOCKSTREAM_API_URL
-            }
-
-    override fun build(): OtsResolver =
-        OtsResolver(
-            explorer =
-                OkHttpBitcoinExplorer(
-                    baseUrl = {
-                        getAPI(usingTor = isTorActive(OkHttpBitcoinExplorer.MEMPOOL_API_URL))
-                    },
-                    client = okHttpClient(OkHttpBitcoinExplorer.MEMPOOL_API_URL),
-                    cache = cache,
-                ),
-            calendar = OkHttpCalendar(okHttpClient),
-        )
-}
+// Re-export from commons for backward compatibility
+typealias TorAwareOkHttpOtsResolverBuilder = com.vitorpamplona.amethyst.commons.model.nip03Timestamp.TorAwareOkHttpOtsResolverBuilder
