+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/HiddenUsersState.kt
@@ -20,103 +20,5 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists
 
-import com.vitorpamplona.amethyst.commons.model.LiveHiddenUsers
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.service.checkNotInMainThread
-import com.vitorpamplona.quartz.nip01Core.core.HexKey
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.MuteTag
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.UserTag
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.WordTag
-import com.vitorpamplona.quartz.utils.DualCase
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combineTransform
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.update
-
-class HiddenUsersState(
-    val muteList: StateFlow<List<MuteTag>>,
-    val blockList: StateFlow<List<MuteTag>>,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    var transientHiddenUsers: MutableStateFlow<Set<String>> = MutableStateFlow(setOf())
-
-    fun assembleLiveHiddenUsers(
-        blockList: List<MuteTag>,
-        muteList: List<MuteTag>,
-        transientHiddenUsers: Set<String>,
-        showSensitiveContent: Boolean?,
-        maxHashtagLimit: Int,
-    ): LiveHiddenUsers {
-        val hiddenUsers = blockList.mapNotNullTo(mutableSetOf()) { if (it is UserTag) it.pubKey else null } + muteList.mapNotNull { if (it is UserTag) it.pubKey else null }
-        val hiddenWords = blockList.mapNotNullTo(mutableSetOf()) { if (it is WordTag) it.word else null } + muteList.mapNotNull { if (it is WordTag) it.word else null }
-
-        return LiveHiddenUsers(
-            showSensitiveContent = showSensitiveContent,
-            hiddenWordsCase = hiddenWords.map { DualCase(it.lowercase(), it.uppercase()) },
-            hiddenUsersHashCodes = hiddenUsers.mapTo(HashSet()) { it.hashCode() },
-            spammersHashCodes = transientHiddenUsers.mapTo(HashSet()) { it.hashCode() },
-            hiddenUsers = hiddenUsers,
-            spammers = transientHiddenUsers,
-            hiddenWords = hiddenWords,
-            maxHashtagLimit = maxHashtagLimit,
-        )
-    }
-
-    val flow: StateFlow<LiveHiddenUsers> =
-        combineTransform(
-            blockList,
-            muteList,
-            transientHiddenUsers,
-            settings.syncedSettings.security.showSensitiveContent,
-            settings.syncedSettings.security.maxHashtagLimit,
-        ) { blockList, muteList, transientHiddenUsers, showSensitiveContent, maxHashtagLimit ->
-            checkNotInMainThread()
-            emit(assembleLiveHiddenUsers(blockList, muteList, transientHiddenUsers, showSensitiveContent, maxHashtagLimit))
-        }.onStart {
-            emit(
-                assembleLiveHiddenUsers(
-                    blockList.value,
-                    muteList.value,
-                    transientHiddenUsers.value,
-                    settings.syncedSettings.security.showSensitiveContent.value,
-                    settings.syncedSettings.security.maxHashtagLimit.value,
-                ),
-            )
-        }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                LiveHiddenUsers(
-                    showSensitiveContent = null,
-                    hiddenWordsCase = emptyList(),
-                    hiddenUsersHashCodes = emptySet(),
-                    spammersHashCodes = emptySet(),
-                    hiddenUsers = emptySet(),
-                    spammers = emptySet(),
-                    hiddenWords = emptySet(),
-                ),
-            )
-
-    fun resetTransientUsers() {
-        transientHiddenUsers.update {
-            emptySet()
-        }
-    }
-
-    fun showUser(pubkeyHex: HexKey) {
-        transientHiddenUsers.update { it - pubkeyHex }
-    }
-
-    fun hideUser(pubkeyHex: HexKey) {
-        transientHiddenUsers.update { it + pubkeyHex }
-    }
-
-    fun isHidden(pubkeyHex: HexKey) = pubkeyHex in transientHiddenUsers.value
-}
+// Re-export from commons for backward compatibility
+typealias HiddenUsersState = com.vitorpamplona.amethyst.commons.model.nip51Lists.HiddenUsersState
